tests/wpt/tests/css/css-gaps/parsing/rule-outset-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Gap Decorations: rule-outset sets longhands</title>
<link rel="help" href="https://www.w3.org/TR/css-gaps-1/#propdef-rule-outset">
<meta name="assert" content="rule-outset supports the full grammar '<length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
const rule_properties = {
    'rule-outset': ['column-rule-outset',
                    'row-rule-outset'],
};

const testCases = [
    {
        input: '0px',
        expected: {
            column: '0px',
            row: '0px'
        }
    },
    {
        input: '10px',
        expected: {
            column: '10px',
            row: '10px'
        }
    },
    {
        input: '50%',
        expected: {
            column: '50%',
            row: '50%'
        }
    },
  ];

for(rule_property in rule_properties) {
    const [column, row] = rule_properties[rule_property];

    for (const { input, expected } of testCases) {
            test_shorthand_value(rule_property, input, {
                [column]: expected.column,
                [row]: expected.row
            });
    }
}
</script>
</body>
</html>
